+++ src/data/types/amino_acids/view_traits.rs
@@ -1,8 +1,9 @@
 use crate::{
-    data::views::{impl_len, impl_restrict_for_wrapper, impl_views_for_wrapper},
+    data::views::{impl_len, impl_restrict_for_wrapper, impl_slice_for_wrapper, impl_views_for_wrapper},
     prelude::{AminoAcids, AminoAcidsView, AminoAcidsViewMut},
 };
 
 impl_len!(AminoAcids, AminoAcidsView, AminoAcidsViewMut, 0);
 impl_views_for_wrapper!(AminoAcids, AminoAcidsView, AminoAcidsViewMut);
+impl_slice_for_wrapper!(AminoAcids, AminoAcidsView, AminoAcidsViewMut);
 impl_restrict_for_wrapper!(AminoAcids, AminoAcidsView, AminoAcidsViewMut);